+++ src/main/java/net/detalk/api/service/SessionAuthService.java
@@ -13,8 +13,8 @@ import net.detalk.api.repository.MemberExternalRepository;
 import net.detalk.api.repository.MemberProfileRepository;
 import net.detalk.api.repository.MemberRepository;
 import net.detalk.api.support.TimeHolder;
-import net.detalk.api.support.security.OAuthProvider;
-import net.detalk.api.support.security.OAuthUser;
+import net.detalk.api.support.security.oauth.OAuthProvider;
+import net.detalk.api.support.security.oauth.OAuthUser;
 import net.detalk.api.support.security.SecurityRole;
 import net.detalk.api.support.util.StringUtil;
 import org.springframework.security.core.authority.AuthorityUtils;
