+++ src/main/java/net/detalk/api/support/security/SecurityConfig.java
@@ -6,6 +6,9 @@ import lombok.RequiredArgsConstructor;
 import net.detalk.api.service.AuthService;
 import net.detalk.api.support.error.ErrorCode;
 import net.detalk.api.support.error.ErrorMessage;
+import net.detalk.api.support.security.oauth.JwtOAuthSuccessHandler;
+import net.detalk.api.support.security.oauth.OAuth2AuthorizationRequestRepository;
+import net.detalk.api.support.security.oauth.OAuthFailHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
