+++ src/main/java/net/detalk/api/support/security/oauth/JwtOAuthSuccessHandler.java
@@ -1,4 +1,4 @@
-package net.detalk.api.support.security;
+package net.detalk.api.support.security.oauth;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -22,7 +22,7 @@ import java.util.Arrays;
 
 import static net.detalk.api.support.Constant.COOKIE_ACCESS_TOKEN;
 import static net.detalk.api.support.Constant.COOKIE_REFRESH_TOKEN;
-import static net.detalk.api.support.security.OAuth2AuthorizationRequestRepository.REDIRECT_URI_COOKIE_NAME;
+import static net.detalk.api.support.security.oauth.OAuth2AuthorizationRequestRepository.REDIRECT_URI_COOKIE_NAME;
 
 @Slf4j
 @Component
