+++ src/main/java/net/detalk/api/support/security/oauth/OAuthProvider.java
@@ -1,4 +1,4 @@
-package net.detalk.api.support.security;
+package net.detalk.api.support.security.oauth;
 
 import lombok.Getter;
